BAEL - 2251
Moved the code from spring-boot-mvc to spring-all
