+++ spring-all/src/main/java/org/baeldung/caching/eviction/CacheEvictionMain.java
@@ -1,0 +1,17 @@
+package org.baeldung.caching.eviction;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+
+@SpringBootApplication
+@EnableCaching
+@EnableScheduling
+public class CacheEvictionMain {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CacheEvictionMain.class, args);
+    }
+}
