+++ spring-all/src/main/java/org/baeldung/caching/eviction/cotrollers/CachingController.java
@@ -1,0 +1,18 @@
+package org.baeldung.caching.eviction.cotrollers;
+
+import org.baeldung.caching.eviction.service.CachingService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CachingController {
+
+    @Autowired
+    CachingService cachingService;
+
+    @GetMapping("clearAllCaches")
+    public void clearAllCaches() {
+        cachingService.evictAllCaches();
+    }
+}
