+++ spring-boot-mvc/src/main/java/com/baeldung/springbootmvc/SpringBootMvcApplication.java
@@ -2,12 +2,8 @@ package com.baeldung.springbootmvc;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
-@EnableCaching
-@EnableScheduling
 public class SpringBootMvcApplication {
 
     public static void main(String[] args) {
