--- spring-boot-mvc/src/main/java/com/baeldung/springbootmvc/caching/CachingController.java
@@ -1,17 +1,0 @@
-package com.baeldung.springbootmvc.caching;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class CachingController {
-	
-	@Autowired
-	CachingService cachingService;
-	
-	@GetMapping("clearAllCaches")
-	public void clearAllCaches() {
-		cachingService.evictAllCaches();
-	}
-}
